refactor: logger refactor
