+++ src/scratch/utils/logging/console.py
@@ -1,0 +1,5 @@
+"""Shared rich console."""
+
+from rich.console import Console
+
+console = Console()
